* master:
  changelog: add version reset [v0.3.3]
  archutils: proper includes and #ifdef handling
  logsystem: fix logger level setting from config
  config: use proper variable type for size
  connectordata: fix connectortype output operator<<
  run: add --reset function (looking at you, mlang)
  config|bvs: clarify module specification
  module: simplify definition
  bvs: improve main documentation
  config: fix command line usage example
  control: adapt to thread naming changes
  archutils: change nameThisThread signature to use std::string, prefix with 'bvs:'
  archutils: move includes where they belong
  cmake: fix BVS_THREAD_NAMES message
  logger: prevent compiler warnings about unused stuff if log system disabled
  bvs: add version info

+++ lib/src/config.cc
@@ -149,7 +149,7 @@ Config& Config::loadConfigFile(const std::string& configFile)
 		lineNumber++;
 
 		//remove all unquoted whitespace/tabs, strip inline comments
-		for (unsigned int i=0; i<line.length(); i++)
+		for (size_t i=0; i<line.length(); i++)
 		{
 			if (line[i]=='#') break;
 			if (line[i]=='\'' || line [i]=='"')
